backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.24.3
pandas==2.1.4
python-dotenv==1.0.0
supabase==2.0.2
urllib3==2.1.0
beautifulsoup4==4.12.2
dnspython==2.4.2
tldextract==5.1.1
python-whois==0.8.0
cryptography==41.0.8 
asyncpg 